+++ tee-worker/app-libs/stf/src/trusted_call_litentry.rs
@@ -261,7 +261,7 @@ impl TrustedCallSigned {
 		bn: ParentchainBlockNumber,
 		hash: H256,
 	) -> StfResult<()> {
-		debug!("build assertion, who {:?}, assertion {:?}", account_id_to_string(&who), assertion);
+		debug!("request vc, who {:?}, assertion {:?}", account_id_to_string(&who), assertion);
 		ensure_enclave_signer_account(&enclave_account)?;
 		let id_graph = ita_sgx_runtime::pallet_imt::Pallet::<Runtime>::get_id_graph(&who);
 		let mut vec_identity: BoundedVec<Identity, MaxIdentityLength> = vec![].try_into().unwrap();
